third_party/blink/web_tests/platform/generic/storage/indexeddb/setversion-not-blocked-expected.txt

Test that setVersion is not blocked if handle closed in versionchange handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "setversion-not-blocked.html"
indexedDB.deleteDatabase(dbname)
versionchangeEventFired = false
indexedDB.open(dbname)


prepareDatabase():
h1 = event.target.result
h1.onversionchange = h1OnVersionChange
indexedDB.open(dbname, 2)


h1OnVersionChange():
versionchangeEventFired = true
PASS event.oldVersion is 1
PASS event.newVersion is 2
h1.close()


h2UpgradeNeeded():
PASS event.oldVersion is 1
PASS event.newVersion is 2


h2OpenSuccess():
PASS versionchangeEventFired is true
PASS successfullyParsed is true

TEST COMPLETE
